completed homepage publish script

+++ site/bin/publish-home.sh
@@ -19,6 +19,8 @@
 # under the License.
 #
 
+cd `dirname $0`/..
+
 USERNAME=$1
 
 if [ "${USERNAME}" == "" ]; then
@@ -27,13 +29,43 @@ if [ "${USERNAME}" == "" ]; then
   exit 1
 fi
 
+if [ ! -d ../target/site/home ]; then
+  bin/generate-home.sh || exit 1
+  echo
+fi
+
 read -s -p "Password for SVN user ${USERNAME}: " PASSWORD
 echo
 
 SVN_CMD="svn --no-auth-cache --username=${USERNAME} --password=${PASSWORD}"
 
+cd ..
 rm -rf target/svn
-bin/generate-home.sh
-
 mkdir -p target/svn
-${SVN_CMD} co --depth immediates https://svn.apache.org/repos/asf/tinkerpop/site target/svn
+
+${SVN_CMD} co --depth files https://svn.apache.org/repos/asf/tinkerpop/site target/svn
+
+cd target/svn
+
+find ../site/home -mindepth 1 -maxdepth 1 -type d | xargs -n1 basename | xargs -r ${SVN_CMD} update
+
+diff -rq ./ ../site/home/ | awk -f ../../bin/publish-docs.awk | sed 's/^\(.\) \//\1 /g' > ../publish-home.files
+
+for file in $(cat ../publish-home.files | awk '/^[AU]/ {print $2}')
+do
+  if [ -d "../site/home/${file}" ]; then
+    mkdir -p "${file}" && cp -r "../site/home/${file}"/* "$_"
+  else
+    mkdir -p "`dirname ${file}`" && cp "../site/home/${file}" "$_"
+  fi
+done
+
+cat ../publish-home.files | awk '/^A/ {print $2}' | xargs -r svn add --parents
+cat ../publish-home.files | awk '/^D/ {print $2}' | xargs -r svn delete
+
+CHANGES=$(cat ../publish-home.files | wc -l)
+
+if [ ${CHANGES} -gt 0 ]; then
+  ${SVN_CMD} commit -m "Deploy TinkerPop homepage"
+fi
+
